fintrack-bot/requirements.txt
python-telegram-bot[job-queue]>=20.0
anthropic
supabase
python-dotenv
openpyxl
pdfplumber
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
